Add Sonatype publishing to build.gradle.kts and update Java version
The build.gradle.kts has been updated to include code for publishing packages to Sonatype. A new plugin, "com.gradleup.nmcp", was added for this purpose. The Java version for Github Actions has also been updated from '17' to '21'. A number of parameters for publication in the Sonatype Nexus Repository Manager such as developer details and licensing were also included in the update.